Add package-lock.json to version control
It's considered best practice to commit the `package-lock.json` file to source control, as this will help to ensure that all installs use the same versions of all dependencies. This can avoid things breaking if a new version of a dependency of sub-dependency introduces a bug or a breaking change.